+++ libcontainer/userns/userns_linux_test.go
@@ -1,6 +1,6 @@
 // +build linux
 
-package system
+package userns
 
 import (
 	"strings"
